+++ packages/fr_acdd/lib/src/model/extracted_api_schema.dart
@@ -1,0 +1,11 @@
+class ExtractedApiSchema {
+  const ExtractedApiSchema({
+    required this.suggestedPath,
+    required this.description,
+    this.explicitPath = false,
+  });
+
+  final String suggestedPath;
+  final String description;
+  final bool explicitPath;
+}